Remove first implementation, make the last one the final one
