+++ packages/visualizations-react/src/index.tsx
@@ -2,5 +2,4 @@ export type { Props } from './components/Props';
 export { default as Chart } from './components/Chart';
 export { default as MarkdownText } from './components/MarkdownText';
 export { default as KpiCard } from './components/KpiCard';
-export { default as Map } from './components/Map';
 export { default as Choropleth } from './components/Choropleth';
