+++ packages/visualizations/src/components/Map/MapRender.svelte
