+++ packages/visualizations/src/index.ts
@@ -1,7 +1,6 @@
 export { default as Chart, _ChartJs, _ChartDataLabels } from './components/Chart';
 export { default as MarkdownText } from './components/MarkdownText';
 export { default as KpiCard } from './components/KpiCard';
-export { default as Map } from './components/Map';
-export { default as Choropleth } from './components/Map2';
+export { default as Choropleth } from './components/Map';
 export * from './types';
 export * from './components/types';
